feat: Remove expanded indicator of asset new value
